Transfer transformStickyEndData test and make stickyEnds return cloned list of objects
